Restructured testing, added a check for cuda availability, fixed firstIdx bug
